Move to new `plugin_development` backend
