+++ src/python/pants/backend/python/macros/pants_requirement_caof.py
@@ -50,9 +50,10 @@ class PantsRequirementCAOF:
             "2.10.0.dev0",
             "the `pants_requirement` macro",
             (
-                "Use the target `pants_requirements` instead. Specifically, we recommend "
-                "deleting all `pants_requirement` calls and replacing them with a single "
-                "`pants_requirements(name='pants')`.\n\n"
+                "Use the target `pants_requirements` instead. First, add "
+                "`pants.backend.plugin_development` to `[GLOBAL].backend_packages` in "
+                "`pants.toml`. Then, delete all `pants_requirement` calls and replace them with "
+                "a single `pants_requirements(name='pants')`.\n\n"
                 "By default, `pants_requirements` will generate a `python_requirement` target for "
                 "both `pantsbuild.pants` and `pantsbuild.pants.testutil`.\n\n"
                 "The address for the generated targets will be different, e.g. "
