+++ src/python/pants/backend/python/register.py
@@ -19,7 +19,6 @@ from pants.backend.python.goals import (
     setup_py,
     tailor,
 )
-from pants.backend.python.macros import pants_requirements
 from pants.backend.python.macros.pants_requirement_caof import PantsRequirementCAOF
 from pants.backend.python.macros.pipenv_requirements_caof import PipenvRequirementsCAOF
 from pants.backend.python.macros.poetry_requirements_caof import PoetryRequirementsCAOF
@@ -85,8 +84,6 @@ def rules():
         *setuptools.rules(),
         *tailor.rules(),
         *target_types_rules.rules(),
-        # Macros
-        *pants_requirements.rules(),
     )
 
 
@@ -101,6 +98,4 @@ def target_types():
         PythonTestsGeneratorTarget,
         PythonTestTarget,
         PythonTestUtilsGeneratorTarget,
-        # Macros
-        pants_requirements.PantsRequirementsTargetGenerator,
     ]
